requirements.txt
Django==4.2.4
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
psycopg2==2.9.3
psycopg2-binary==2.9.7